# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Exam Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions
      parameters:
        - name: examId
          in: query
          schema:
            type: integer
        - name: sectionId
          in: query
          schema:
            type: integer
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: topicId
          in: query
          schema:
            type: integer
        - name: difficulty
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get question by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # TOPIC MOCKS
...
  /attempts:
    get:
      operationId: listAttempts
      tags: [attempts]
      summary: List my attempts
      parameters:
        - name: examId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attempt"
    post:
      operationId: startAttempt
      tags: [attempts]
      summary: Start an attempt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttemptInput"
      responses:
        "201":
          description: Started attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptDetail"
  /attempts/{id}:
    get:
      operationId: getAttempt
      tags: [attempts]
      summary: Get attempt (resume)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attempt detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptDetail"
    patch:
      operationId: updateAttempt
      tags: [attempts]
      summary: Save attempt progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttemptUpdate"
      responses:
        "200":
          description: Updated attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attempt"
  /attempts/{id}/submit:
    post:
      operationId: submitAttempt
      tags: [attempts]
      summary: Submit attempt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"

  # RESULTS
...
components:
  schemas:
...
    Section:
      type: object
      required: [id, name, examId, duration, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        examId:
          type: integer
        duration:
          type: integer
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Question:
      type: object
      required: [id, questionText, optionA, optionB, optionC, optionD, correctAnswer, marks, createdAt]
      properties:
        id:
          type: integer
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
        explanation:
          type: ["string", "null"]
        marks:
          type: number
        negativeMarks:
          type: number
        difficulty:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        examId:
          type: ["integer", "null"]
        sectionId:
          type: ["integer", "null"]
        subjectId:
          type: ["integer", "null"]
        topicId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    QuestionInput:
      type: object
      required: [questionText, optionA, optionB, optionC, optionD, correctAnswer, marks]
      properties:
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        marks:
          type: number
        negativeMarks:
          type: number
        difficulty:
          type: string
        year:
          type: integer
        examId:
          type: integer
        sectionId:
          type: integer
        subjectId:
          type: integer
        topicId:
          type: integer
    QuestionUpdate:
      type: object
      properties:
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        marks:
          type: number
        negativeMarks:
          type: number
        difficulty:
          type: string
        year:
          type: integer
        examId:
          type: integer
        sectionId:
          type: integer
        subjectId:
          type: integer
        topicId:
          type: integer

    # TOPIC MOCKS
...
    Attempt:
      type: object
      required: [id, userId, entityType, entityId, status, startedAt, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        entityType:
          type: string
          enum: [EXAM, QUIZ, TOPIC_MOCK]
        entityId:
          type: integer
        status:
          type: string
          enum: [IN_PROGRESS, SUBMITTED, AUTO_SUBMITTED]
        currentSection:
          type: ["integer", "null"]
        timeRemaining:
          type: ["integer", "null"]
        startedAt:
          type: string
          format: date-time
        submittedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        examTitle:
          type: ["string", "null"]
    AttemptDetail:
      type: object
      required: [id, userId, entityType, entityId, status, startedAt, questions]
      properties:
        id:
          type: integer
        userId:
          type: integer
        entityType:
          type: string
        entityId:
          type: integer
        status:
          type: string
        currentSection:
          type: ["integer", "null"]
        timeRemaining:
          type: ["integer", "null"]
        startedAt:
          type: string
          format: date-time
        submittedAt:
          type: ["string", "null"]
          format: date-time
        answers:
          type: array
          items:
            $ref: "#/components/schemas/Answer"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        sections:
          type: array
          items:
            $ref: "#/components/schemas/Section"
        examTitle:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        hasSectionalTimer:
          type: boolean
    AttemptInput:
      type: object
      required: [entityType, entityId]
      properties:
        entityType:
          type: string
          enum: [EXAM, QUIZ, TOPIC_MOCK]
        entityId:
          type: integer
    AttemptUpdate:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerInput"
        currentSection:
          type: integer
        timeRemaining:
          type: integer
    Answer:
      type: object
      required: [questionId]
      properties:
        questionId:
          type: integer
        selectedOption:
          type: ["string", "null"]
        markedForReview:
          type: boolean
        timeSpent:
          type: ["integer", "null"]
    AnswerInput:
      type: object
      required: [questionId]
      properties:
        questionId:
          type: integer
        selectedOption:
          type: ["string", "null"]
        markedForReview:
          type: boolean
        timeSpent:
          type: integer

    # RESULTS
    Result:
      type: object
      required: [id, attemptId, userId, totalScore, maxScore, correctCount, wrongCount, skippedCount, accuracy, createdAt]
      properties:
        id:
          type: integer
        attemptId:
          type: integer
        userId:
          type: integer
        totalScore:
          type: number
        maxScore:
          type: number
        correctCount:
          type: integer
        wrongCount:
          type: integer
        skippedCount:
          type: integer
        accuracy:
          type: number
        timeTaken:
          type: ["integer", "null"]
        rank:
          type: ["integer", "null"]
        percentile:
          type: ["number", "null"]
        examTitle:
          type: ["string", "null"]
        entityType:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...