# lib/api-zod/src/generated/types/attemptInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Exam Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AttemptInputEntityType } from './attemptInputEntityType';

export interface AttemptInput {
  entityType: AttemptInputEntityType;
  entityId: number;
}