# lib/api-zod/src/generated/types/questionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Exam Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionUpdate {
  questionText?: string;
  optionA?: string;
  optionB?: string;
  optionC?: string;
  optionD?: string;
  correctAnswer?: string;
  explanation?: string;
  marks?: number;
  negativeMarks?: number;
  difficulty?: string;
  year?: number;
  examId?: number;
  sectionId?: number;
  subjectId?: number;
  topicId?: number;
}